gpflux/layers/trackable_layer.py
#
# Copyright (c) 2021 The GPflux Contributors.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
"""Utility layer that tracks variables in :class:`tf.Module`."""


import itertools
from functools import wraps
from typing import Any, Callable, Optional, Sequence

import tensorflow as tf


class TrackableLayer(tf.keras.layers.Layer):
    pass
